# ledger-app/openapi.yaml
openapi: 3.1.0
info:
  description: |-
    E-wallet API on a double-entry ledger: open wallets, deposit, transfer between wallets, read balances and entry history.

    Conventions:

    - **Amounts** are strings holding an integer in the minor unit of the currency (`"150000"` is 150,000 VND), so no client rounds them. Entry amounts are signed from the wallet's side: negative is money out, positive is money in.
    - **Ids** of wallets and transfers are UUIDs.
    - **Requests that move money** need an `Idempotency-Key` header. Retrying with the same key never moves money twice.
    - **Errors** are RFC 9457 Problem Details with the media type `application/problem+json`. The `type` member tells them apart.
    - **There is no authentication.** Whoever can reach the API can act on every wallet.
  title: Ledgerly API
  version: "1"
servers:
- description: Local
  url: http://localhost:8080
...
paths:
  /v1/admin/deposits:
    post:
      description: |-
        Credits the wallet and debits the system funding account, which is allowed to go negative, in one database transaction. This is how wallets get money in development, tests and demos: no bank is involved, and since the API has no authentication the endpoint must not be reachable from outside.

        The call needs an `Idempotency-Key`, with the same rules as a transfer: a retry with the same key returns the stored response with `Idempotent-Replayed: true` and credits nothing again. The response has no `Location`: a deposit cannot be read back on its own, it appears as an entry of the wallet.
      operationId: createDeposit
      parameters:
      - description: "Identifies this operation, so that it is safe to send again.\
          \ Generate one value for each operation (a UUID is recommended) and send\
          \ that same value on every retry of it. 1 to 64 visible ASCII characters.\
          \ The key and the response stored under it are kept for at least 24 hours."
        example: 9b2e4c1a-7d3f-4a6b-8c5d-0e1f2a3b4c5d
        in: header
        name: Idempotency-Key
        required: true
        schema:
          type: string
          pattern: "[\\x21-\\x7E]{1,64}"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositResponse"
          description: The deposit is complete. Also the answer when a completed deposit
            is retried with the same Idempotency-Key.
          headers:
            Idempotent-Replayed:
              description: "Present with the value `true` when this is the stored\
                \ response of an earlier request with the same Idempotency-Key, and\
                \ nothing was done again."
              schema:
                type: string
                enum:
                - "true"
              style: simple
        "400":
          content:
            application/problem+json:
              examples:
                validation-error:
                  description: "The body, a parameter or a header is missing or malformed.\
                    \ Nothing was done, and an Idempotency-Key sent with the request\
                    \ was not used up."
                  summary: Request is not valid
                  value:
                    status: 400
                    title: Request is not valid
                    type: /problems/validation-error
              schema:
                $ref: "#/components/schemas/Problem"
          description: Request is not valid
        "404":
          content:
            application/problem+json:
              examples:
                wallet-not-found:
                  description: No wallet has this id. System accounts are not wallets.
                  summary: Wallet not found
                  value:
                    status: 404
                    title: Wallet not found
                    type: /problems/wallet-not-found
              schema:
                $ref: "#/components/schemas/Problem"
          description: Wallet not found
        "409":
          content:
            application/problem+json:
              examples:
                idempotency-in-progress:
                  description: "An earlier request with the same Idempotency-Key has\
                    \ not finished. Wait for the time in Retry-After, then send the\
                    \ request again with the same key."
                  summary: Request with this Idempotency-Key is still in progress
                  value:
                    status: 409
                    title: Request with this Idempotency-Key is still in progress
                    type: /problems/idempotency-in-progress
              schema:
                $ref: "#/components/schemas/Problem"
          description: Request with this Idempotency-Key is still in progress
          headers:
            Retry-After:
              description: Seconds to wait before sending the same request again.
              schema:
                type: integer
                format: int32
                example: 1
        "422":
          content:
            application/problem+json:
              examples:
                currency-mismatch:
                  description: The currency of the request is not the currency of
                    a wallet it names.
                  summary: Currency does not match the wallet
                  value:
                    status: 422
                    title: Currency does not match the wallet
                    type: /problems/currency-mismatch
                idempotency-key-reused:
                  description: "The Idempotency-Key was first sent with another method,\
                    \ path or body. Use a new key for every operation."
                  summary: Idempotency-Key was already used with a different request
                  value:
                    status: 422
                    title: Idempotency-Key was already used with a different request
                    type: /problems/idempotency-key-reused
              schema:
                $ref: "#/components/schemas/Problem"
          description: "Idempotency-Key was already used with a different request,\
            \ or: Currency does not match the wallet"
        "500":
          content:
            application/problem+json:
              examples:
                internal-error:
                  description: "Something unexpected failed. The cause is in the server\
                    \ log, not in the response. A request that moves money was rolled\
                    \ back and may be retried with the same Idempotency-Key."
                  summary: Internal error
                  value:
                    status: 500
                    title: Internal error
                    type: /problems/internal-error
              schema:
                $ref: "#/components/schemas/Problem"
          description: Internal error
        "503":
          content:
            application/problem+json:
              examples:
                overloaded:
                  description: "The request could not get the locks it needs in time.\
                    \ Nothing was done. Wait for the time in Retry-After, then send\
                    \ the request again with the same Idempotency-Key."
                  summary: Service is overloaded
                  value:
                    status: 503
                    title: Service is overloaded
                    type: /problems/overloaded
              schema:
                $ref: "#/components/schemas/Problem"
          description: Service is overloaded
          headers:
            Retry-After:
              description: Seconds to wait before sending the same request again.
              schema:
                type: integer
                format: int32
                example: 1
      summary: Deposit money into a wallet
      tags:
      - Deposits
...
components:
  schemas:
    DepositRequest:
      type: object
      description: The deposit to make.
      properties:
        amount:
          type: string
          description: "Amount to credit, as a string holding a positive integer in\
            \ the minor unit of the currency. No sign, no leading zeros, no decimals,\
            \ at most 18 digits."
          example: "500000"
          pattern: "[1-9][0-9]{0,17}"
        currency:
          type: string
          description: ISO 4217 code of the currency. Must be the currency of the
            wallet. Only VND is supported.
          example: VND
        walletId:
          type: string
          format: uuid
          description: Wallet to credit.
          example: 01a11ad9-7eb8-7b3e-9d1f-2f3c6f1c2a10
      required:
      - amount
      - currency
      - walletId
    DepositResponse:
      type: object
      description: A completed deposit.
      properties:
        amount:
          type: string
          description: "Amount credited, as a string holding a positive integer in\
            \ the minor unit of the currency."
          example: "500000"
        createdAt:
          type: string
          format: date-time
          description: When the deposit was committed.
        currency:
          type: string
          description: ISO 4217 code of the currency.
          example: VND
        id:
          type: string
          format: uuid
          description: Id of the deposit. Also the `transactionId` of the entry it
            wrote to the wallet.
          example: 01a11ad9-7f10-7459-a273-deb24dbb4296
        status:
          type: string
          description: "Always `COMPLETED`: a deposit that exists has credited the\
            \ wallet."
          enum:
          - COMPLETED
        walletId:
          type: string
          format: uuid
          description: Wallet that was credited.
          example: 01a11ad9-7eb8-7b3e-9d1f-2f3c6f1c2a10
...
    Problem:
      type: object
      description: "An error, as RFC 9457 Problem Details. Every error of the API\
        \ has this shape. Tell errors apart by `type`, never by `title` or `detail`."
      properties:
        detail:
          type: string
          description: "What went wrong in this occurrence, for a human to read."
          example: "Wallet 01a11ad9-7eb8-7b3e-9d1f-2f3c6f1c2a10 has 350000 VND, the\
            \ transfer needs 9999999 VND"
        errors:
          type: array
          description: "Only on `validation-error`, and only when the rejected inputs\
            \ are known: one item per rejected input."
          items:
            type: object
            properties:
              field:
                type: string
                description: "The JSON field, query parameter or header that was rejected."
                example: amount
              message:
                type: string
                description: Why it was rejected.
                example: must be a positive integer of minor units with at most 18
                  digits
        instance:
          type: string
          description: Path of the request that failed.
          example: /v1/transfers
        status:
          type: integer
          format: int32
          description: The HTTP status of the response.
          example: 422
        title:
          type: string
          description: "Short name of the kind of problem, the same for every occurrence."
          example: Insufficient funds
        type:
          type: string
          description: "Identifies the kind of problem. Stable: safe to branch on.\
            \ Errors raised by the HTTP layer itself, such as an unknown path or a\
            \ method that is not allowed, have no `type`."
          enum:
          - /problems/validation-error
          - /problems/wallet-not-found
          - /problems/transfer-not-found
          - /problems/idempotency-in-progress
          - /problems/idempotency-key-reused
          - /problems/insufficient-funds
          - /problems/same-account-transfer
          - /problems/currency-mismatch
          - /problems/unsupported-currency
          - /problems/internal-error
          - /problems/overloaded
          example: /problems/insufficient-funds
...